Moved backend auth tests to their own folder, and updated the test run call. Frontend: Fixed an issue where Navbar reference still existed in App.js. Dashboard: Added dark mode designs
